+++ docqueries/dijkstra/doc-pgr_dijkstraCost.test.sql
@@ -33,4 +33,23 @@ SELECT * FROM pgr_dijkstraCost(
         'select id, source, target, cost, reverse_cost from edge_table',
             ARRAY[5, 3, 4], ARRAY[5, 3, 4]);
 
+-- pgr_dijkstra combinations SQL
+-------------------------------------------------------------------------------
+
 \echo --q8
+
+SELECT * FROM pgr_dijkstraCost(
+    'SELECT id, source, target, cost, reverse_cost FROM edge_table',
+    'SELECT source, target FROM combinations_table',
+    FALSE
+);
+
+\echo --q9
+
+SELECT * FROM pgr_dijkstraCost(
+    'SELECT id, source, target, cost FROM edge_table',
+    'SELECT * FROM (VALUES (2, 3), (2, 5), (11, 3), (11, 5)) AS combinations (source, target)',
+    FALSE
+);
+
+\echo --q10