+++ include/drivers/dijkstra/dijkstra_driver.h
@@ -10,6 +10,10 @@ Function's developer:
 Copyright (c) 2015 Celia Virginia Vergara Castillo
 Mail: vicky_vergara@hotmail.com
 
+Copyright (c) 2020 The combinations_sql signature is added by Mahmoud SAKR
+and Esteban ZIMANYI
+mail: m_attia_sakr@yahoo.com, estebanzimanyi@gmail.com
+
 ------
 
 This program is free software; you can redistribute it and/or modify
@@ -39,6 +43,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #endif
 
 #include "c_types/pgr_edge_t.h"
+#include "c_types/pgr_combination_t.h"
 #include "c_types/general_path_element_t.h"
 
 
@@ -70,6 +75,26 @@ extern "C" {
             char** notice_msg,
             char** err_msg);
 
+
+    //  CREATE OR REPLACE FUNCTION pgr_dijkstra(
+    //  sql text,
+    //  combinations_sql text,
+    //  directed boolean default true,
+    void do_pgr_combinations_dijkstra(
+            pgr_edge_t  *data_edges,
+            size_t total_tuples,
+            pgr_combination_t  *combinations,
+            size_t total_combinations,
+            bool directed,
+            bool only_cost,
+            bool normal,
+
+            General_path_element_t **return_tuples,
+            size_t *return_count,
+
+            char** log_msg,
+            char** notice_msg,
+            char** err_msg);
 #ifdef __cplusplus
     }
 #endif