+++ sql/scripts/build-extension-update-files1.pl
@@ -21,10 +21,11 @@ my $version_2_4 = qr/(2.4.[\d+])/;
 my $version_2_5 = qr/(2.5.[\d+])/;
 my $version_2_6 = qr/(2.6.[\d+])/;
 my $version_3_0 = qr/(3.0.[\d+])/;
+my $version_3_1 = qr/(3.1.[\d+])/;
 my $version_2 = qr/(2.[\d+].[\d+])/;
 my $version_3 = qr/(3.[\d+].[\d+])/;
 
-my $current = $version_3_0;
+my $current = $version_3_1;
 
 
 sub Usage {
@@ -48,8 +49,8 @@ my $curr_sql_file_name = "$output_directory/pgrouting--$version.sql";
 ######################################################
 
 # Verify Currently accepted old and new versions
-die "ERROR: 'build-extension-update-files1.pl' expected old version: 2.0.x ~~ 2.5.x\nHINT: Please check:\n  file: sql/CMakeLists.txt\n  Section: 'TARGET: update_files'"
-    unless $old_version =~ /$version_2_0|$version_2_1|$version_2_2|$version_2_3|$version_2_4|$version_2_5|$version_2_6|$version_3_0/;
+die "ERROR: 'build-extension-update-files1.pl' expected old version: 2.0.x ~~ 2.6.x\nHINT: Please check:\n  file: sql/CMakeLists.txt\n  Section: 'TARGET: update_files'"
+    unless $old_version =~ /$version_2_0|$version_2_1|$version_2_2|$version_2_3|$version_2_4|$version_2_5|$version_2_6|$version_3_0|$version_3_1/;
 
 die "ERROR: 'build-extension-update-files1.pl' expected version: 3.0.x\nHINT: ERROR on PGROUTING_VERSION variable\nor File sql/scripts/build-extension-update-files1.pl needs to be reviewed"
     unless $version =~ /$current/;