+++ tools/release-scripts/check-release.sh
@@ -52,12 +52,12 @@ fi
 
 
 MAYOR=3
-MINOR=0
-MICRO=2
+MINOR=1
+MICRO=0
 RC=""
-PREV_REL=3.0.0
+PREV_REL=3.0.2
 PREV_RELS="
-    3.0.1 3.0.0
+    3.0.2 3.0.1 3.0.0
     2.6.3 2.6.2 2.6.1 2.6.0
     "
 # These releases are not for update